requirements.txt
subprocess.run==0.0.8
python-nmap==0.6.1
openpyxl==3.0.6